+++ indico/modules/events/tracks/controllers.py
@@ -25,7 +25,6 @@ from sqlalchemy.orm import subqueryload
 from indico.core.db.sqlalchemy.descriptions import RENDER_MODE_WRAPPER_MAP
 from indico.legacy.pdfinterface.conference import ProgrammeToPDF
 from indico.modules.events.controllers.base import RHDisplayEventBase
-from indico.modules.events.layout.util import get_menu_entry_by_name
 from indico.modules.events.management.controllers import RHManageEventBase
 from indico.modules.events.tracks.forms import ProgramForm, TrackForm
 from indico.modules.events.tracks.models.tracks import Track
@@ -122,7 +121,6 @@ class RHDeleteTrack(RHManageTrackBase):
 
 class RHDisplayTracks(RHDisplayEventBase):
     def _process(self):
-        page_title = get_menu_entry_by_name('program', self.event).localized_title
         program = track_settings.get(self.event, 'program')
         render_mode = track_settings.get(self.event, 'program_render_mode')
         program = RENDER_MODE_WRAPPER_MAP[render_mode](program)
@@ -131,8 +129,7 @@ class RHDisplayTracks(RHDisplayEventBase):
                            subqueryload('abstract_reviewers'))
                   .order_by(Track.position)
                   .all())
-        return WPDisplayTracks.render_template('display.html', self.event,
-                                               page_title=page_title, program=program, tracks=tracks)
+        return WPDisplayTracks.render_template('display.html', self.event, program=program, tracks=tracks)
 
 
 class RHTracksPDF(RHDisplayEventBase):